requirements.txt
discord.py
python-dotenv
tinydb